Separate nominal types from utility functions into different modules
Split nominal.ts into two focused modules:
- nominal-types.ts: Pure type definitions only
- nominal-utils.ts: Constructor functions and utilities

Update all import statements across the codebase to use the new modules.
This improves code organization and separation of concerns.

Address PR comment #4 - separate types from functions/logic.
